Move `get_devices_with_keys_by_user` to `EndToEndKeyWorkerStore`
this seems a better fit for it.

This commit simply moves the existing code: no other changes at all.
